// src/types/employee.ts
// ===============================
// Address
// ===============================

export interface Address {
  city: string;
  state: string;
  country: string;
}

// ===============================
// Employee
// ===============================

export interface Employee {
  id: number;

  name: string;

  email: string;

  department: string;

  role: string;

  salary: number;

  joinDate: string;

  isActive: boolean;

  skills: string[];

  address: Address;

  projects: number;

  lastReview: string;

  performanceRating: number;
}